extract some bridge functions into their own preload.ts file
