+++ libs/coin-modules/coin-cosmos/src/bridge/index.ts
@@ -5,19 +5,14 @@ import {
   makeSync,
 } from "@ledgerhq/coin-framework/bridge/jsHelpers";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
-import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets";
-import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import type { AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
 
-import { CosmosValidatorsManager } from "../CosmosValidatorsManager";
 import { CosmosAPI } from "../api/Cosmos";
-import cryptoFactory from "../chain/chain";
-import { CoinConfig, getCoinConfig, setCoinConfig } from "../config";
+import { CoinConfig, setCoinConfig } from "../config";
 import { createTransaction } from "../createTransaction";
 import { estimateMaxSpendable } from "../estimateMaxSpendable";
 import getTransactionStatus from "../getTransactionStatus";
 import resolver from "../hw-getAddress";
-import { asSafeCosmosPreloadData, setCosmosPreloadData } from "../preloadedData";
 import { prepareTransaction } from "../prepareTransaction";
 import {
   assignFromAccountRaw,
@@ -29,13 +24,12 @@ import { buildSignOperation } from "../signOperation";
 import { getAccountShape } from "../synchronisation";
 import type {
   CosmosAccount,
-  CosmosCurrencyConfig,
-  CosmosValidatorItem,
   Transaction,
   TransactionStatus,
 } from "../types";
 import { CosmosSigner } from "../types/signer";
 import { updateTransaction } from "../updateTransaction";
+import { getPreloadStrategy, hydrate, preload } from "./preload";
 
 const sync = makeSync({ getAccountShape });
 
@@ -47,45 +41,10 @@ function buildCurrencyBridge(signerContext: SignerContext<CosmosSigner>): Curren
     getAddressFn: getAddressWrapper(getAddress),
   });
 
-  const getPreloadStrategy = () => ({
-    preloadMaxAge: 30 * 1000,
-  });
-
   return {
     getPreloadStrategy,
-    preload: async (currency: CryptoCurrency) => {
-      const config = getCoinConfig(currency);
-      const cosmosValidatorsManager = new CosmosValidatorsManager(
-        getCryptoCurrencyById(currency.id),
-        { endPoint: (config as unknown as CosmosCurrencyConfig).lcd },
-      );
-      const validators = await cosmosValidatorsManager.getValidators();
-      setCosmosPreloadData(currency.id, {
-        validators,
-      });
-
-      return Promise.resolve({
-        validators,
-        config,
-      });
-    },
-    hydrate: (
-      data: { validators?: CosmosValidatorItem[]; config: CosmosCurrencyConfig },
-      currency: CryptoCurrency,
-    ) => {
-      if (!data || typeof data !== "object") return;
-      const relatedImpl = cryptoFactory(currency.id);
-      relatedImpl.lcd = data.config.lcd;
-      relatedImpl.minGasPrice = data.config.minGasPrice;
-      relatedImpl.ledgerValidator = data.config?.ledgerValidator;
-      const { validators } = data;
-      if (!validators || typeof validators !== "object" || !Array.isArray(validators)) return;
-      const cosmosValidatorsManager = new CosmosValidatorsManager(
-        getCryptoCurrencyById(currency.id),
-      );
-      cosmosValidatorsManager.hydrateValidators(validators);
-      setCosmosPreloadData(currency.id, asSafeCosmosPreloadData(data));
-    },
+    preload,
+    hydrate,
     scanAccounts,
   };
 }
@@ -94,7 +53,6 @@ function buildAccountBridge(
   signerContext: SignerContext<CosmosSigner>,
 ): AccountBridge<Transaction, CosmosAccount, TransactionStatus> {
   const getAddress = resolver(signerContext);
-  // const getAddress = signerGetAddress(signerContext);
 
   const receive = makeAccountBridgeReceive(getAddressWrapper(getAddress));
   const signOperation = buildSignOperation(signerContext);
